+++ src/components/tmField/index.js
@@ -1,0 +1,5 @@
+import tmField from "./tmField.vue"
+
+export default function install(Vue) {
+  Vue.component("tm-field", tmField)
+}