+++ src/components/tmHelloWorld/index.js
@@ -1,0 +1,5 @@
+import tmHelloWorld from "./tmHelloWorld.vue"
+
+export default function install(Vue) {
+  Vue.component("tm-hello-world", tmHelloWorld)
+}