+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-* Added `./src/index.js` for easy importing
+* Added `./src/index.js` to support `import`
 * Added a README and CHANGELOG
 * Added a placeholder HelloWorld component